Delta Time passing and Memory Leak Patch
Delta time and Total time of the engine is now available in all update methods for proper speed scaling. The reason the camera was slow was due to my PC at home being a beast compared to the lab machines (6-7x the FPS).
Memory leaks patched.
Title bar stats fixed.